Run validate on form start.
There is probably a better way to do this. The form should have `invalid` set to true initially to disable the submit button until all errors are corrected. Error visibility can be controlled by `touched && error`. My assumption is that the form can still be submitted with warnings, although I don't know the difference between the two. Is there an acknowledgement step for warnings? Or is it purely just an informational chunk of state you can use to show additional information based on user input?