Run validation on every value change
Remove the options to run validation on blur or change events, and instead run validation on every value change.

The benefit is that the validation state will _always_ be current for the given value state. This also greatly simplifies the internal validation state management, doing away with the `stale` flags entirely.

Assuming that the provided validation callback is not expensive, this should not negatively affect performance.