Fix SyntheticEvent detection in minified builds
Fixes #7

Comparing the name of the constructor proved to be a terrible idea; it fails completely when function names are mangled by a JS minifier.

Any approach to detecting `SyntheticEvent` instances based on property inspection will have a small chance of false matches, however a raw value returned from an input component to containing `stopPropagation` and `preventDefault` functions is a very unlikely case.